requirements.txt
websocket-client
pandas
numpy
streamlit
statsmodels
plotly
sqlalchemy